Create reusable ConfirmInfo component
